[Bugfix] HTCCL: calibrate the pipe chunk in fp16, not in emulated bf16
The pipe-chunk sweep runs a real all_reduce over a synthetic payload and the
size it picks governs every later transfer. The payload was bfloat16.

bf16 has no hardware conversion on a Turing (sm75) rank or a gfx900 rank -- the
two cards this transport exists for. The kernel is conversion-only (to_f /
from_f, no __nv_bfloat16 arithmetic), so it compiles and runs there, which is
why this was silent: the sweep completes and reports a plausible number. What
it times is emulated per-element conversion on a payload whose dtype those
ranks never carry -- they run fp16, because sm75 and gfx900 have no bf16 at all
and the loader forces the fallback.

fp16 is the honest probe on every card here, not just on the two: same 2 bytes
per element, so the bandwidth being measured is identical, and a single
hardware conversion instruction on cards with native bf16 as well.

This is the only bf16 tensor an sm75 rank allocates in the cross-vendor
configuration; the transport is otherwise dtype-agnostic.

Test (CPU, no device, collectives stubbed):
test/registered/unit/distributed/test_htccl_calibration_dtype.py
Red before / green after on the dtype; the "a chunk size is still chosen"
half was green throughout. test_htccl_port.py and test_htccl_transport_seam.py
unchanged (38 passed together).

+++ python/sglang/srt/distributed/device_communicators/htccl_device.py
@@ -889,14 +889,21 @@ class HTCCLDeviceTransport:
 
         import torch.distributed as dist
 
-        # Collective sweep over the REAL all_reduce path (16 MiB bf16
+        # Collective sweep over the REAL all_reduce path (16 MiB fp16
         # payload). All ranks iterate the candidates in the same order
         # and decide on the summed times, so every rank picks the same
         # value — the per-chunk flag protocol requires agreement, hence
         # pipe_mib is a scalar, not per-rank.
-        payload = torch.randn(
-            8 * 1024 * 1024, dtype=torch.bfloat16, device=self.device
-        )
+        #
+        # fp16 and NOT bf16: both are 2 bytes, so the bandwidth measured is
+        # the same, but bf16 has no hardware conversion on a Turing (sm75) or
+        # gfx900 rank -- the two cards this transport exists for. The kernel's
+        # to_f/from_f would be emulated for the probe while the traffic those
+        # ranks actually carry is fp16, so the sweep would time arithmetic
+        # that never runs and pick a chunk size for it. fp16 is converted by a
+        # single instruction on every card this transport targets, including
+        # the ones that do have native bf16.
+        payload = torch.randn(8 * 1024 * 1024, dtype=torch.float16, device=self.device)
         candidates = [1, 2, 4, 8]
         my_times = []
         for mib in candidates:
